--- Projects/Toolinder/Feature/MyPage/Interface/Sources/Main/Components/Cell/GoalSummaryCellStore.swift
@@ -1,8 +1,0 @@
-//
-//  GoalSummaryCellStore.swift
-//  DyingFeatureGoalDemo
-//
-//  Created by 송영모 on 2023/08/08.
-//
-
-import Foundation
